+++ vision_agent/tools/tools_types.py
